Convert AS_IF mapping expressions through the member's own type
MSVC C4244: the AS_IF macros assigned the conversion expression directly,
so an int64_t or double result narrowed into an int, Evpu, or Efix member at
every such row. assignFrom now accepts any value type and both AS_IF macros
route through it, matching the plain mapping rows. The pre-3.5 stem unit
constant converts its double explicitly.

+++ src/import/options/stem_options.cpp
@@ -122,7 +122,7 @@ void insertRecoveredConnection(const musx::dom::DocumentPtr& document, const std
 
 // One staff position is one half space, so the pre-3.5 lengths convert through musxdom's own
 // constant rather than through a 12 written here.
-constexpr int evpuPerStaffPosition = musx::dom::EVPU_PER_STAFF_POSITION;
+constexpr int evpuPerStaffPosition = static_cast<int>(musx::dom::EVPU_PER_STAFF_POSITION);
 
 // Finale 3.5 and later, and the DCL epoch entire.
 const FieldMapping stemScalarFields[] = {
